Created Font class
font stores a set of texture names and the appropriate data to render a
string using the characters represented by those textures. This class is
currently the active rendering method for strings, but does not function
properly. creating a single character from make_tex_from_ttf() sort of
works, but is ot active.